+++ CHANGELOG.md
@@ -28,6 +28,8 @@
 
 ### jannovar-core
 
+* `VariantList` implements the `List<Annotation>` interface now and the
+  `entries` member has become private.
 * Adding `VariantEffect#SEQUENCE_VARIANT` for variants with unknown
   effects.
 * `GenomeChange.toString()` now always converts to forward strand.